Updated TempSensor Reading
- Added in a new way to read temp sensor values.  Compare the previous temp with the current temp and see if it is greater than a predefined value.  If it is, then do not update the temperature.  This will help better compensate for large temp fluctations within 1 second.  The temperature should not change drastically (over 10 degrees, even over 5 degrees) within 1 second especially in a fish tank environment.
- Added in ability to revert back to previous temp method of always storing the value read in by the sensor.  Use #define DirectTempSensor to always have the temp updated.
- Removed a few extra LCD.Clears of a single pixel while storing the temp parameters in memory.
- Trimmed down Developer Mode string on version page to just be Dev Mode.
- Minor code cleanup to save memory

+++ ReefAngel_TempSensor/ReefAngel_TempSensor.cpp
@@ -78,7 +78,7 @@ int ReefAngel_TempSensorClass::ReadTemperature(byte addr[8], byte unit)
 			ds.reset();
 			ds.select(addr);
 			ds.write(0xBE);
-			for (int i = 0; i < 2; i++)
+			for (byte i = 0; i < 2; i++)
 			{	     // we need 9 bytes
 				data[i] = ds.read();
 			}
@@ -88,7 +88,7 @@ int ReefAngel_TempSensorClass::ReadTemperature(byte addr[8], byte unit)
 				Temp=(data[1]<<8)+data[0];//take the two bytes from the response relating to temperature
 				Temp=Temp/1.6;
 				if (Temp == 0) return 0;
-				//if (unit==0) Temp=Temp*1.8+320; // comment this line out to get celcius
+
 				if ( unit == DEGREE_F )
                     Temp = Temp * 1.8 + 320;
 	//		}
